Move abstract entry model to zinnia.models_bases.entry
Fixes #219
